Refactoring of classes and methods
