+++ app/src/main/java/com/m/androidNativeApp/Client.java
@@ -1,14 +1,11 @@
 package com.m.androidNativeApp;
 
-import android.content.Context;
-
 import com.apollographql.apollo.ApolloClient;
-import com.m.helpper.MobileService;
+import com.m.helper.MobileService;
 
 import okhttp3.OkHttpClient;
 
 public class Client {
-    private static String SERVER_URL;
     private MobileService mobileService = MobileService.getInstance();
 
     public ApolloClient setupApollo() {
