+++ app/src/main/java/com/m/helper/DTO/config/AndroidConfig.java
@@ -1,4 +1,4 @@
-package com.m.helpper.DTO.config;
+package com.m.helper.DTO.config;
 
 public class AndroidConfig {
     private String variantId;
