+++ app/src/main/java/com/m/helper/DTO/config/SyncConfig.java
@@ -1,4 +1,4 @@
-package com.m.helpper.DTO.config;
+package com.m.helper.DTO.config;
 
 public class SyncConfig {
     private String websocketUrl;
