+++ app/src/main/java/com/m/helper/MobileService.java
@@ -1,4 +1,4 @@
-package com.m.helpper;
+package com.m.helper;
 
 import android.content.Context;
 import android.content.res.AssetManager;
@@ -8,12 +8,12 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.m.helpper.DTO.Keycloak;
-import com.m.helpper.DTO.MobileServiceJSON;
-import com.m.helpper.DTO.Push;
-import com.m.helpper.DTO.Service;
-import com.m.helpper.DTO.SyncApp;
-import com.m.helpper.DTO.config.AndroidConfig;
+import com.m.helper.DTO.Keycloak;
+import com.m.helper.DTO.MobileServiceJSON;
+import com.m.helper.DTO.Push;
+import com.m.helper.DTO.Service;
+import com.m.helper.DTO.SyncApp;
+import com.m.helper.DTO.config.AndroidConfig;
 
 
 import java.io.BufferedReader;
@@ -22,15 +22,16 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 
-
-public class MobileService implements MobileServices {
+public class MobileService implements IMobileService {
 
     private static MobileService INSTANCE = null;
     private MobileServiceJSON config = new MobileServiceJSON();
-
     private AssetManager assetManager;
-
     private Gson gson = new Gson();
+
+    private final String SYNC_APP = "sync-app";
+    private final String KEYCLOAK = "keycloak";
+    private final String PUSH = "push";
 
     private MobileService(Context context){
         System.out.println("APP : Setting up Mobile Services");
@@ -56,65 +57,52 @@ public class MobileService implements MobileServices {
 
     @Override
     public String getGraphqlServer() {
-        SyncApp service = (SyncApp) config.getServiceByType("sync-app");
+        SyncApp service = (SyncApp) config.getServiceByType(SYNC_APP);
         return service.getUrl();
     }
 
     @Override
     public String getKIssuer() {
-        Keycloak service = (Keycloak) config.getServiceByType("keycloak");
+        Keycloak service = (Keycloak) config.getServiceByType(KEYCLOAK);
         String server = service.getConfig().getAuth_server_url();
         String realm = service.getConfig().getRealm();
-        String kIssuer = server + "/realms/" + realm + "/";
-        return kIssuer;
+        return server + "/realms/" + realm + "/";
     }
 
     @Override
     public String getKClientId() {
-        Keycloak service = (Keycloak) config.getServiceByType("keycloak");
+        Keycloak service = (Keycloak) config.getServiceByType(KEYCLOAK);
         return service.getConfig().getResource();
     }
 
     @Override
     public String getPushUrl() {
-        Push service = (Push) config.getServiceByType("push");
+        Push service = (Push) config.getServiceByType(PUSH);
         return service.getUrl();
     }
 
     @Override
     public String getPushVariantId() {
-        Push service = (Push) config.getServiceByType("push");
+        Push service = (Push) config.getServiceByType(PUSH);
         return service.getConfig().getVariantId();
     }
 
     @Override
     public String getPushVariantSecret() {
-        Push service = (Push) config.getServiceByType("push");
+        Push service = (Push) config.getServiceByType(PUSH);
         return service.getConfig().getVariantSecret();
     }
 
     private void readMobileServiceJSON(){
         try {
-            InputStream IS = assetManager.open("config/mobile-services.json");
-            InputStreamReader ISR = new InputStreamReader(IS);
-            BufferedReader BR = new BufferedReader(ISR);
-
-            JsonParser parser = new JsonParser();
-            JsonObject array = parser.parse(BR).getAsJsonObject();
+            BufferedReader bufferedReader = getBufferReader();
+            JsonObject array = getJsonObject(bufferedReader);
 
             config.setClientId(array.get("clientId").getAsString());
             config.setNamespace(array.get("namespace").getAsString());
 
             JsonArray services = array.get("services").getAsJsonArray();
-
-            for (int i = 0; i < services.size(); i++){
-                Service newService = buildService(services.get(i).getAsJsonObject());
-                if (newService != null){
-                    config.addService(newService);
-                } else {
-                    System.out.println("APP: Service was null");
-                }
-            }
+            configureServices(services);
 
         } catch (FileNotFoundException e) {
             System.out.println("APP: FileNotFoundException " + e.toString());
@@ -127,15 +115,14 @@ public class MobileService implements MobileServices {
 
     private Service buildService(JsonObject service){
         Service result = null;
-//        TODO add enums for all string lookup values
         switch (service.get("type").getAsString()) {
-            case "sync-app":
+            case SYNC_APP:
                 result = gson.fromJson(service, SyncApp.class);
                 break;
-            case "keycloak":
+            case KEYCLOAK:
                 result = gson.fromJson(service, Keycloak.class);
                 break;
-            case "push":
+            case PUSH:
                 Push r = gson.fromJson(service, Push.class);
                 JsonElement jsonConfig = service.get("config");
                 JsonElement jsonAndroidConfig = jsonConfig.getAsJsonObject().get("android");
@@ -147,9 +134,29 @@ public class MobileService implements MobileServices {
                 System.out.println("APP: No type found");
 
         }
-
-
         return result;
     }
 
+    private BufferedReader getBufferReader() throws IOException {
+        InputStream IS = assetManager.open("config/mobile-services.json");
+        InputStreamReader ISR = new InputStreamReader(IS);
+        return new BufferedReader(ISR);
+    }
+
+    private JsonObject getJsonObject(BufferedReader bufferedReader){
+        JsonParser parser = new JsonParser();
+        return parser.parse(bufferedReader).getAsJsonObject();
+    }
+
+    private void configureServices(JsonArray services) {
+        for (int i = 0; i < services.size(); i++){
+            Service newService = buildService(services.get(i).getAsJsonObject());
+            if (newService != null){
+                config.addService(newService);
+            } else {
+                System.out.println("APP: Service was null");
+            }
+        }
+
+    }
 }
